Add useIngredientTags hook
